Renamed ChangeColorTable to ChangeColorTableWindow
