+++ Gm1KonverterCrossPlatform/Views/ChangeColorTableWindow.axaml
@@ -5,7 +5,7 @@
         xmlns:controls="clr-namespace:ThemeEditor.Controls.ColorPicker;assembly=ThemeEditor.Controls.ColorPicker"
         mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
         Width="800" Height="400"
-        x:Class="Gm1KonverterCrossPlatform.Views.ChangeColorTable"
+        x:Class="Gm1KonverterCrossPlatform.Views.ChangeColorTableWindow"
         Title="{DynamicResource ChangeColorPalette}"
         WindowStartupLocation="CenterOwner">
