+++ Gm1KonverterCrossPlatform/Views/ChangeColorTableWindowDialogBox.axaml
@@ -5,7 +5,7 @@
         xmlns:controls="clr-namespace:ThemeEditor.Controls.ColorPicker;assembly=ThemeEditor.Controls.ColorPicker"
         mc:Ignorable="d" d:DesignWidth="300" d:DesignHeight="100"
         Width="300" Height="100"
-        x:Class="Gm1KonverterCrossPlatform.Views.ChangeColorTableDialogBox"
+        x:Class="Gm1KonverterCrossPlatform.Views.ChangeColorTableWindowDialogBox"
         Title="{DynamicResource ChangeColorPalette}"
         WindowStartupLocation="CenterOwner">
